refactor: Move settings out of `OrtModel`
Move the settings from the `OrtModel` to the `WorkbenchController`.
These are global settings so they should be handled on the app level
instead of the level of a loaded ORT result file.
