+++ src/main/kotlin/ui/WorkbenchController.kt
@@ -1,15 +1,60 @@
 package org.ossreviewtoolkit.workbench.ui
 
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
+import com.fasterxml.jackson.dataformat.yaml.YAMLGenerator
+import com.fasterxml.jackson.dataformat.yaml.YAMLMapper
+import com.fasterxml.jackson.module.kotlin.readValue
+import com.fasterxml.jackson.module.kotlin.registerKotlinModule
+
 import java.nio.file.Path
 
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.NonCancellable
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
+import kotlinx.coroutines.withContext
+
+import org.ossreviewtoolkit.utils.common.safeMkdirs
+import org.ossreviewtoolkit.utils.ort.ortDataDirectory
 import org.ossreviewtoolkit.workbench.model.OrtApiState
 import org.ossreviewtoolkit.workbench.model.OrtModel
+import org.ossreviewtoolkit.workbench.model.WorkbenchSettings
 import org.ossreviewtoolkit.workbench.state.DialogState
 
+private const val ORT_WORKBENCH_CONFIG_DIRNAME = "workbench"
+private const val ORT_WORKBENCH_CONFIG_FILENAME = "settings.yml"
+
 class WorkbenchController {
-    val ortModel = OrtModel()
+    companion object {
+        private val settingsMapper =
+            YAMLMapper(YAMLFactory().disable(YAMLGenerator.Feature.WRITE_DOC_START_MARKER)).apply {
+                registerKotlinModule()
+            }
+    }
+
+    private val scope = CoroutineScope(Dispatchers.Default)
+    private val mutex = Mutex()
+
+    private val settingsFile =
+        ortDataDirectory.resolve(ORT_WORKBENCH_CONFIG_DIRNAME).resolve(ORT_WORKBENCH_CONFIG_FILENAME)
+
+    private val _settings = MutableStateFlow(WorkbenchSettings.default())
+    val settings: StateFlow<WorkbenchSettings> = _settings
+
+    private val _error = MutableStateFlow<String?>(null)
+    val error: StateFlow<String?> = _error
+
+    val ortModel = OrtModel(settings)
 
     val openResultDialog = DialogState<Path?>()
+
+    init {
+        scope.launch { loadSettings() }
+    }
 
     suspend fun openOrtResult() {
         if (ortModel.state.value !in listOf(OrtApiState.LOADING_RESULT, OrtApiState.PROCESSING_RESULT)) {
@@ -17,6 +62,35 @@ class WorkbenchController {
             if (path != null) {
                 ortModel.loadOrtResult(path.toFile())
             }
+        }
+    }
+
+    suspend fun updateSettings(settings: WorkbenchSettings) {
+        mutex.withLock {
+            withContext(Dispatchers.IO + NonCancellable) {
+                _settings.value = settings
+                saveSettings(settings)
+            }
+        }
+    }
+
+    private fun loadSettings() {
+        val settings =
+            settingsFile.takeIf { it.isFile }?.let { settingsMapper.readValue(it) } ?: WorkbenchSettings.default()
+
+        _settings.value = settings
+
+        if (!settingsFile.exists()) {
+            saveSettings(settings)
+        }
+    }
+
+    private fun saveSettings(settings: WorkbenchSettings) {
+        runCatching {
+            settingsFile.parentFile.safeMkdirs()
+            settingsMapper.writeValue(settingsFile, settings)
+        }.onFailure {
+            _error.value = "Could not save settings at ${settingsFile.absolutePath}: ${it.message}"
         }
     }
 }
